refactor: replace face_recognition with insightface and remove unused test file
