+++ vision/packages/vision_general/scripts/face_recognition_node.py
@@ -1,18 +1,10 @@
 #!/usr/bin/env python3
-
-"""
-Node to recognize known people and
-name new faces.
-- Publisher for largest face detected
-coordinates for arm following.
-- Service to assign name to face.
-"""
-
+# ros2 run vision_general face_recognition_node.py
 import os
 import pathlib
 
 import cv2
-import face_recognition
+import insightface
 import numpy as np
 import rclpy
 import rclpy.duration
@@ -20,17 +12,19 @@ import tqdm
 from ament_index_python.packages import get_package_share_directory
 from cv_bridge import CvBridge
 from geometry_msgs.msg import Point
+from insightface.app import FaceAnalysis
 from rclpy.node import Node
 from sensor_msgs.msg import CameraInfo, Image
 from std_msgs.msg import Bool, String
 
 from frida_constants.vision_constants import (
     CAMERA_INFO_TOPIC,
+    CAMERA_TOPIC,
     DEPTH_IMAGE_TOPIC,
     FACE_RECOGNITION_IMAGE,
+    FLIP_IMAGE_TOPIC,
     FOLLOW_BY_TOPIC,
     FOLLOW_TOPIC,
-    IMAGE_ORIENTED_TOPIC,
     PERSON_LIST_TOPIC,
     PERSON_NAME_TOPIC,
     SAVE_NAME_TOPIC,
@@ -41,7 +35,7 @@ from frida_interfaces.srv import SaveName
 
 DEFAULT_NAME = "ale"
 TRACK_THRESHOLD = 50
-MATCH_THRESHOLD = 0.5
+MATCH_THRESHOLD = 0.35
 MAX_DEGREE = 1
 RESIZE_FACTOR = 1
 
@@ -49,6 +43,24 @@ PATH = str(pathlib.Path(__file__).parent)
 PATH = get_package_share_directory("vision_general")
 KNOWN_FACES_PATH = PATH + "/Utils/known_faces"
 
+INSIGHTFACE_MODEL = "buffalo_l"
+INSIGHTFACE_CTX_ID = 0
+
+def _cosine_similarity(a: np.ndarray, b: np.ndarray) -> float:
+    """Return cosine similarity in [−1, 1]. Higher = more similar."""
+    a = a / (np.linalg.norm(a) + 1e-10)
+    b = b / (np.linalg.norm(b) + 1e-10)
+    return float(np.dot(a, b))
+
+
+def _cosine_distances(known: list[np.ndarray], query: np.ndarray) -> np.ndarray:
+    """Return 1 − cosine_similarity for each embedding in *known*."""
+    if not known:
+        return np.array([])
+    mat = np.stack(known)                       
+    mat = mat / (np.linalg.norm(mat, axis=1, keepdims=True) + 1e-10)
+    q   = query / (np.linalg.norm(query) + 1e-10)
+    return 1.0 - mat @ q                        
 
 class FaceRecognition(Node):
     def __init__(self):
@@ -64,7 +76,7 @@ class FaceRecognition(Node):
 
         self.image_subscriber = self.create_subscription(
             Image,
-            IMAGE_ORIENTED_TOPIC,
+            CAMERA_TOPIC,
             self.image_callback,
             self._img_qos,
             callback_group=self.callback_group,
@@ -96,17 +108,18 @@ class FaceRecognition(Node):
             self.follow_by_name_callback,
             callback_group=self.callback_group,
         )
-        self.follow_publisher = self.create_publisher(Point, FOLLOW_TOPIC, 10)
-        self.view_pub = self.create_publisher(Image, FACE_RECOGNITION_IMAGE, 10)
-        self.name_publisher = self.create_publisher(String, PERSON_NAME_TOPIC, 10)
-        self.person_list_publisher = self.create_publisher(
-            PersonList, PERSON_LIST_TOPIC, 10
-        )
+
+        self.follow_publisher      = self.create_publisher(Point,      FOLLOW_TOPIC,           10)
+        self.view_pub              = self.create_publisher(Image,      FACE_RECOGNITION_IMAGE, 10)
+        self.name_publisher        = self.create_publisher(String,     PERSON_NAME_TOPIC,      10)
+        self.person_list_publisher = self.create_publisher(PersonList, PERSON_LIST_TOPIC,      10)
 
         self.verbose = self.declare_parameter("verbose", True)
         self.annotated_frame = []
-        self.vision_active = True
-        self.is_processing = False
+        self.vision_active   = True
+        self.is_processing   = False
+        self.flip_image      = False
+
         self.create_subscription(
             Bool,
             "/vision/face_recognition/active",
@@ -114,56 +127,71 @@ class FaceRecognition(Node):
             10,
             callback_group=self.callback_group,
         )
+        self.create_subscription(
+            Bool,
+            FLIP_IMAGE_TOPIC,
+            self._flip_callback,
+            10,
+            callback_group=self.callback_group,
+        )
+
         self.setup()
         self.create_timer(0.2, self.run, callback_group=self.callback_group)
-        # self.create_timer(0.05, self.publish_image)
+
 
     def setup(self):
-        """Setup face recognition, reset variables and load models"""
-        random = face_recognition.load_image_file(f"{KNOWN_FACES_PATH}/random.png")
-        random_encodings = face_recognition.face_encodings(random)[0]
+        """Initialise InsightFace pipeline and load known-face embeddings."""
+
+        self.get_logger().info(
+            f"Loading InsightFace model '{INSIGHTFACE_MODEL}' on ctx_id={INSIGHTFACE_CTX_ID} …"
+        )
+        self.app = FaceAnalysis(
+            name=INSIGHTFACE_MODEL,
+            providers=["CUDAExecutionProvider", "CPUExecutionProvider"],
+        )
+        self.app.prepare(ctx_id=INSIGHTFACE_CTX_ID, det_size=(640, 640))
         self.pbar.update(1)
-        self.new_name = ""
-        self.image_view = None
-        self.image = None
-        self.prev_faces = []
-        self.curr_faces = []
-        self.depth_image = []
-        self.follow_name = "area"
-        self.id = None
-        self.processing_id = rclpy.duration.Infinite
+
+        self.new_name        = ""
+        self.image_view      = None
+        self.image           = None
+        self.prev_faces      = []
+        self.curr_faces      = []
+        self.depth_image     = []
+        self.follow_name     = "area"
+        self.id              = None
+        self.processing_id   = rclpy.duration.Infinite
+
         self.default_name = self.declare_parameter("default_name", DEFAULT_NAME)
         self.default_name = self.default_name.value
 
-        self.people = [[random_encodings, "random"]]
-        self.people_encodings = [random_encodings]
-        self.people_names = ["random"]
+        self.people_encodings: list[np.ndarray] = []
+        self.people_names:     list[str]        = []
+        self.people_encodings.append(np.zeros(512, dtype=np.float32))
+        self.people_names.append("random")
 
         self.clear()
         self.process_imgs()
-        self.get_logger().info("Face Recognition Ready")
+        self.pbar.update(1)
+        self.get_logger().info("Face Recognition Ready  (insightface / ArcFace 512-d)")
+
 
     def image_callback(self, data):
-        """Callback to get image from camera"""
-        # self.get_logger().info("img received reentrant")
-        self.id = data.header.stamp
-        # self.get_logger().info(f"s{self.id}")
+        self.id    = data.header.stamp
         self.image = self.bridge.imgmsg_to_cv2(data, "bgr8")
 
     def depth_callback(self, data):
-        """Callback to receive depth image from camera"""
         try:
-            depth_image = self.bridge.imgmsg_to_cv2(data, "32FC1")
-            self.depth_image = depth_image
+            self.depth_image = self.bridge.imgmsg_to_cv2(data, "32FC1")
         except Exception as e:
             print(f"Error: {e}")
 
     def image_info_callback(self, data):
-        """Callback to receive camera info"""
         self.imageInfo = data
 
+
     def new_name_callback(self, req, res):
-        """Callback to add a new face to the known faces"""
+        """Assign a name to the currently largest detected face."""
         self.get_logger().info("Executing service new face")
         self.new_name = req.name
         if len(self.curr_faces) == 0:
@@ -175,39 +203,35 @@ class FaceRecognition(Node):
         return res
 
     def follow_by_name_callback(self, req, res):
-        """Callback to follow face by name or area"""
+        """Set which face name (or 'area') the arm should follow."""
         self.get_logger().info("Executing service follow by")
         self.follow_name = req.name
         if len(self.curr_faces) == 0:
             self.get_logger().info("No face detected")
             res.success = False
         else:
-            self.get_logger().info(f"New name: {self.follow_name}")
+            self.get_logger().info(f"Follow name: {self.follow_name}")
             res.success = True
         return res
 
+
     def success(self, message) -> None:
-        """Print success message"""
         self.get_logger().info(f"\033[92mSUCCESS:\033[0m {message}")
 
     def publish_image(self) -> None:
-        """Publish image with annotations"""
         if len(self.annotated_frame) > 0:
             self.view_pub.publish(
                 self.bridge.cv2_to_imgmsg(self.annotated_frame, "bgr8")
             )
 
     def process_imgs(self) -> None:
-        """Make encodings of known people images"""
-        self.get_logger().info("Processing images")
+        self.get_logger().info("Processing known-face images …")
         for filename in os.listdir(KNOWN_FACES_PATH):
             if filename == ".DS_Store":
                 continue
-
             self.process_img(filename)
 
     def clear(self) -> None:
-        """Clear previous results"""
         for filename in os.listdir(KNOWN_FACES_PATH):
             if (
                 filename == ".DS_Store"
@@ -215,136 +239,88 @@ class FaceRecognition(Node):
                 or filename == self.default_name + ".png"
             ):
                 continue
-
             file_path = os.path.join(KNOWN_FACES_PATH, filename)
             os.remove(file_path)
 
-    def _align_face(self, crop: np.ndarray) -> np.ndarray:
-        """Rotate crop so eyes are horizontal. Input BGR, return RGB rotated."""
+    def apply_clahe(self, bgr_img: np.ndarray) -> np.ndarray:
         try:
-            rgb = cv2.cvtColor(crop, cv2.COLOR_BGR2RGB)
-            lm_list = face_recognition.face_landmarks(rgb)
-            if not lm_list:
-                return rgb
-            lm = lm_list[0]
-            if "left_eye" not in lm or "right_eye" not in lm:
-                return rgb
-            left = np.mean(lm["left_eye"], axis=0)
-            right = np.mean(lm["right_eye"], axis=0)
-            dy = right[1] - left[1]
-            dx = right[0] - left[0]
-            angle = np.degrees(np.arctan2(dy, dx))
-            eyes_center = ((left[0] + right[0]) / 2.0, (left[1] + right[1]) / 2.0)
-            M = cv2.getRotationMatrix2D(eyes_center, -angle, 1.0)
-            rotated = cv2.warpAffine(
-                rgb,
-                M,
-                (rgb.shape[1], rgb.shape[0]),
-                flags=cv2.INTER_CUBIC,
-                borderMode=cv2.BORDER_REPLICATE,
-            )
-            return rotated
-        except Exception:
-            return cv2.cvtColor(crop, cv2.COLOR_BGR2RGB)
-
-    def apply_clahe(self, rgb_img: np.ndarray) -> np.ndarray:
-        """Apply CLAHE on L channel of an RGB image and return RGB."""
-        try:
-            lab = cv2.cvtColor(rgb_img, cv2.COLOR_RGB2LAB)
+            lab = cv2.cvtColor(bgr_img, cv2.COLOR_BGR2LAB)
             l_img, a, b = cv2.split(lab)
             clahe = cv2.createCLAHE(clipLimit=0.5, tileGridSize=(16, 16))
             l_img = clahe.apply(l_img)
             lab = cv2.merge((l_img, a, b))
-            return cv2.cvtColor(lab, cv2.COLOR_LAB2RGB)
+            return cv2.cvtColor(lab, cv2.COLOR_LAB2BGR)
         except Exception:
-            return rgb_img
+            return bgr_img
 
     def process_img(self, filename: str) -> None:
-        """Process image, obtain encodings and add to known people"""
-        img = face_recognition.load_image_file(f"{KNOWN_FACES_PATH}/{filename}")
-
-        # try:
-        #     # convert to BGR for alignment helper which expects BGR input
-        #     img_bgr = cv2.cvtColor(img, cv2.COLOR_RGB2BGR)
-        #     aligned_rgb = self._align_face(img_bgr)
-        #     preprocessed = self.apply_clahe(aligned_rgb)
-        # except Exception:
-        #     preprocessed = img
-
-        cur_encodings = face_recognition.face_encodings(img)
-
-        if len(cur_encodings) == 0:
-            print("no encodings found")
+        img_path = f"{KNOWN_FACES_PATH}/{filename}"
+        img_bgr  = cv2.imread(img_path)
+        if img_bgr is None:
+            self.get_logger().warn(f"Could not load image: {img_path}")
             return
 
-        if len(cur_encodings) > 0:
-            cur_encodings = cur_encodings[0]
+        img_bgr = self.apply_clahe(img_bgr)
+        faces = self.app.get(img_bgr)
+        if len(faces) == 0:
+            self.get_logger().warn(f"No face found in {filename}, skipping.")
+            return
 
-        self.people_encodings.append(cur_encodings)
+        face = max(faces, key=lambda f: _bbox_area(f.bbox))
+        embedding = face.embedding.astype(np.float32)
+
+        self.people_encodings.append(embedding)
         self.people_names.append(filename[:-4])
-        self.people.append([cur_encodings, filename[:-4]])
 
     def save_face(self, name: str, xc: float, yc: float) -> None:
-        """Save face to list and return Person message"""
+        """Append detected face to curr_faces and face_list PersonList."""
         self.curr_faces.append({"x": xc, "y": yc, "name": name})
-        curr_person = Person()
+        curr_person      = Person()
         curr_person.name = name
-        curr_person.x = int((xc - self.center[0]) * MAX_DEGREE / self.center[0])
-        curr_person.y = int((self.center[1] - yc) * MAX_DEGREE / self.center[1])
+        curr_person.x    = int((xc - self.center[0]) * MAX_DEGREE / self.center[0])
+        curr_person.y    = int((self.center[1] - yc) * MAX_DEGREE / self.center[1])
         self.face_list.list.append(curr_person)
 
     def assign_name(
-        self, left: float, top: float, bottom: float, right: float, xc: float, yc: float
+        self,
+        left: float, top: float, bottom: float, right: float,
+        xc: float,   yc: float,
     ) -> None:
-        """Assign name to largest face detected"""
-
-        crop = self.frame[top:bottom, left:right]
-
-        try:
-            aligned_rgb = self._align_face(crop)
-            save_bgr = self.apply_clahe(aligned_rgb)
-        except Exception:
-            save_bgr = crop
-
+        crop     = self.frame[top:bottom, left:right]
         img_name = f"{self.new_name}.png"
         save_path = f"{KNOWN_FACES_PATH}/{img_name}"
-        cv2.imwrite(save_path, save_bgr)
+        cv2.imwrite(save_path, crop)
+
         self.process_img(img_name)
 
-        # Update prev recognitions for tracker
         for i, face in enumerate(self.curr_faces):
             if face["x"] == xc and face["y"] == yc:
-                index = i
+                self.curr_faces[i]["name"]       = self.new_name
+                self.face_list.list[i].name      = self.new_name
+                break
 
-            # cv2.circle(self.annotated_frame, (int(face["x"]), int(face["y"])), 5, (0, 255, 0), -1)
-
-        self.curr_faces[index]["name"] = self.new_name
-        self.face_list.list[index].name = self.new_name
-        self.success(f"{self.new_name} face saved")
-
+        self.success(f"{self.new_name} face enrolled")
         self.new_name = ""
 
-    def publish_follow_face(self, xc: float, yc: float, largest_face_name: str) -> None:
-        """Publish coordinates for arm to follow face"""
-        # Coordinates to follow largest face
-        difx = 0 if xc == 0 else xc - self.center[0]
-        dify = 0 if yc == 0 else self.center[1] - yc
 
+    def publish_follow_face(
+        self, xc: float, yc: float, largest_face_name: str
+    ) -> None:
+        difx   = 0.0 if xc == 0 else xc - self.center[0]
+        dify   = 0.0 if yc == 0 else self.center[1] - yc
         move_x = difx * MAX_DEGREE / self.center[0]
         move_y = dify * MAX_DEGREE / self.center[1]
 
-        target = Point()
-
+        target   = Point()
         target.x = move_x
         target.y = move_y
-
         self.follow_publisher.publish(target)
 
-        person_seen = String()
+        person_seen      = String()
         person_seen.data = largest_face_name
-
         self.name_publisher.publish(person_seen)
         self.person_list_publisher.publish(self.face_list)
+
 
     def _active_callback(self, msg):
         if msg.data == self.vision_active:
@@ -352,22 +328,24 @@ class FaceRecognition(Node):
         self.vision_active = msg.data
         self.get_logger().info(f"Face recognition active: {self.vision_active}")
 
-    def run(self) -> None:
-        """Run face recognition algorithm"""
+    def _flip_callback(self, msg):
+        if msg.data != self.flip_image:
+            self.flip_image = msg.data
+            self.get_logger().info(f"Flip image: {self.flip_image}")
 
+
+    def run(self) -> None:
         if not self.vision_active:
             return
-
         if self.is_processing:
             return
-
         if self.image is None:
             self.get_logger().info("No image")
             return
+
         self.annotated_frame = self.image
 
         if self.id == self.processing_id:
-            # self.get_logger().info("Skipping image")
             return
 
         self.is_processing = True
@@ -376,161 +354,103 @@ class FaceRecognition(Node):
         finally:
             self.is_processing = False
 
+
     def _run_inference(self) -> None:
-        """Actual face recognition inference (called by run with lock)."""
         self.processing_id = self.id
 
         self.frame = self.image
+        if self.flip_image:
+            self.frame = cv2.rotate(self.frame, cv2.ROTATE_180)
         self.annotated_frame = self.frame.copy()
         self.center = [self.frame.shape[1] / 2, self.frame.shape[0] / 2]
+        detected_faces = self.app.get(self.apply_clahe(self.frame))
 
-        # resized_frame = cv2.resize(
-        #     self.frame, (0, 0), fx=RESIZE_FACTOR, fy=RESIZE_FACTOR
-        # )
+        largest_area         = 0
+        follow_face_params   = None
+        largest_area_params  = None
+        largest_face_name    = ""
 
-        # Find all faces using dlib's CUDA-accelerated CNN model (built with CUDA sm_87)
-        # "cnn" uses GPU via dlib CUDA, "hog" would be CPU-only
-        face_locations = face_recognition.face_locations(self.frame, model="cnn")
-        # print("running")
-        # return
+        self.curr_faces  = []
+        self.face_list   = PersonList()
+        detected         = False
 
-        largest_area = 0
-        follow_face_params = None
-        largest_area_params = None
-        largest_face_name = ""
+        for ins_face in detected_faces:
+            x1, y1, x2, y2 = ins_face.bbox.astype(int)
 
-        self.curr_faces = []
-        self.face_list = PersonList()
-        detected = False
-        face_encodings = None
+            left   = max(x1 - TRACK_THRESHOLD, 0)
+            right  = min(x2 + TRACK_THRESHOLD, self.frame.shape[1])
+            top    = max(y1 - TRACK_THRESHOLD, 0)
+            bottom = min(y2 + TRACK_THRESHOLD, self.frame.shape[0])
 
-        # Process each face
-        for i, location in enumerate(face_locations):
-            # Center of current face
-            scale_factor = 1 / RESIZE_FACTOR
-            centerx = (location[3] + (location[1] - location[3]) / 2) * scale_factor
-            centery = (location[0] + (location[2] - location[0]) / 2) * scale_factor
+            centerx = (x1 + x2) / 2.0
+            centery = (y1 + y2) / 2.0
 
-            top, right, bottom, left = [int(i * scale_factor) for i in location]
             name = "Unknown"
-
-            # Extend bbox
-            left = max(left - TRACK_THRESHOLD, 0)
-            right = min(right + TRACK_THRESHOLD, self.frame.shape[1])
-            top = max(0, top - TRACK_THRESHOLD)
-            bottom = min(bottom + TRACK_THRESHOLD, self.frame.shape[0])
-
-            # Tracking
             flag = False
             for prev_face in self.prev_faces:
-                # If the face is within the tracking threshold
-                if (abs(prev_face["x"] - centerx) < TRACK_THRESHOLD) and (
-                    abs(prev_face["y"] - centery) < TRACK_THRESHOLD
+                if (
+                    abs(prev_face["x"] - centerx) < TRACK_THRESHOLD
+                    and abs(prev_face["y"] - centery) < TRACK_THRESHOLD
                 ):
                     name = prev_face["name"]
                     flag = True
                     break
 
-            # If not a tracked face, then it needs to be processed (compare to known faces)
             if not flag:
-                crop = self.frame[
-                    top:bottom, left:right
-                ]  # BGR crop from original resolution
-                if crop.size == 0:
-                    print("crop 0")
-                    continue
+                query_embedding = ins_face.embedding.astype(np.float32)
 
-                aligned_rgb = self._align_face(crop)  # returns RGB
-                preproc_rgb = self.apply_clahe(aligned_rgb)
-                encs = face_recognition.face_encodings(preproc_rgb)
+                if len(self.people_encodings) > 0:
+                    distances       = _cosine_distances(self.people_encodings, query_embedding)
+                    best_match_idx  = int(np.argmin(distances))
+                    best_similarity = 1.0 - float(distances[best_match_idx])
 
-                if len(encs) == 0:
-                    print("no encoding")
-                    # no encoding found for this crop
-                    if face_encodings is None:
-                        face_encodings = face_recognition.face_encodings(
-                            self.frame, face_locations
-                        )
+                    if best_similarity >= MATCH_THRESHOLD and best_match_idx != 0:
+                        name = self.people_names[best_match_idx]
 
-                    face_encoding = face_encodings[i]
-                    # continue
-                else:
-                    face_encoding = encs[0]
-
-                # See if the face is a match for the known face(s)
-                matches = face_recognition.compare_faces(
-                    face_encoding, self.people_encodings, MATCH_THRESHOLD
-                )
-                face_distances = face_recognition.face_distance(
-                    self.people_encodings, face_encoding
-                )
-                best_match_index = np.argmin(face_distances)
-
-                # If it is known, then the name is updated
-                if matches[best_match_index]:
-                    name = self.people_names[best_match_index]
-
-            xc = left + (right - left) / 2
-            yc = top + (bottom - top) / 2
+            xc   = left + (right - left) / 2.0
+            yc   = top  + (bottom - top) / 2.0
             area = (right - left) * (bottom - top)
 
-            # Add face to list
             self.save_face(name, xc, yc)
             detected = True
 
-            # Show results
-            if flag:
-                cv2.rectangle(
-                    self.annotated_frame,
-                    (left, bottom - 35),
-                    (right, bottom),
-                    (255, 0, 0),
-                    cv2.FILLED,
-                )
-                cv2.rectangle(
-                    self.annotated_frame, (left, top), (right, bottom), (255, 0, 0), 2
-                )
-            else:
-                cv2.rectangle(
-                    self.annotated_frame,
-                    (left, bottom - 35),
-                    (right, bottom),
-                    (0, 0, 255),
-                    cv2.FILLED,
-                )
-                cv2.rectangle(
-                    self.annotated_frame, (left, top), (right, bottom), (0, 0, 255), 2
-                )
-
-            font = cv2.FONT_HERSHEY_DUPLEX
+            color = (255, 0, 0) if flag else (0, 0, 255)
+            cv2.rectangle(
+                self.annotated_frame,
+                (left, bottom - 35), (right, bottom),
+                color, cv2.FILLED,
+            )
+            cv2.rectangle(
+                self.annotated_frame,
+                (left, top), (right, bottom),
+                color, 2,
+            )
             cv2.putText(
                 self.annotated_frame,
                 name,
                 (left + 6, bottom - 6),
-                font,
+                cv2.FONT_HERSHEY_DUPLEX,
                 1.0,
                 (255, 255, 255),
                 1,
             )
 
             if area > largest_area:
-                largest_area = area
+                largest_area        = area
                 largest_area_params = [left, top, right, bottom]
-                largest_face_name = name
+                largest_face_name   = name
 
             if self.follow_name == name:
                 follow_face_params = [left, top, right, bottom]
 
-        xc = 0
-        yc = 0
+        xc = 0.0
+        yc = 0.0
 
-        # For the follow face by name or area:
         if largest_area != 0:
             left, top, right, bottom = largest_area_params
-            xc = left + (right - left) / 2
-            yc = top + (bottom - top) / 2
+            xc = left + (right - left) / 2.0
+            yc = top  + (bottom - top) / 2.0
 
-            # Center of the face
             if self.new_name != "":
                 largest_face_name = self.new_name
                 self.assign_name(left, top, bottom, right, xc, yc)
@@ -538,15 +458,14 @@ class FaceRecognition(Node):
         if self.follow_name != "area":
             if follow_face_params is not None:
                 left, top, right, bottom = follow_face_params
-                xc = left + (right - left) / 2
-                yc = top + (bottom - top) / 2
+                xc = left + (right - left) / 2.0
+                yc = top  + (bottom - top) / 2.0
                 largest_face_name = self.follow_name
             else:
                 detected = False
 
         self.prev_faces = self.curr_faces
 
-        # Calculate the joint degrees for the arm to follow the face
         if detected:
             self.publish_follow_face(xc, yc, largest_face_name)
         else:
@@ -554,12 +473,15 @@ class FaceRecognition(Node):
 
         self.publish_image()
 
+def _bbox_area(bbox) -> float:
+    """Return pixel area of an InsightFace bbox [x1, y1, x2, y2]."""
+    x1, y1, x2, y2 = bbox
+    return max(0.0, x2 - x1) * max(0.0, y2 - y1)
 
 def main(args=None):
     rclpy.init(args=args)
-
     try:
-        node = FaceRecognition()
+        node     = FaceRecognition()
         executor = rclpy.executors.MultiThreadedExecutor(5)
         executor.add_node(node)
         executor.spin()
